rename sidebar to navbar and added logger route
