+++ ethernet-view/src/components/Navbar/Navbar.module.scss
@@ -1,6 +1,6 @@
 @use "../../styles/styles.scss";
 
-.sidebarWrapper {
+.navbarWrapper {
     width: fit-content;
     display: flex;
     flex-direction: column;
