+++ ethernet-view/src/components/Navbar/NavbarItem/NavbarItem.module.scss
@@ -11,5 +11,5 @@
 }
 
 .active {
-    color: styles.getColor("orange", 60);
+    filter: brightness(0) saturate(100%) invert(69%) sepia(35%) saturate(2228%) hue-rotate(336deg) brightness(97%) contrast(96%);
 }
