+++ ethernet-view/src/layouts/AppLayout/AppLayout.tsx
@@ -1,14 +1,15 @@
-import { Sidebar } from "components/Sidebar/Sidebar"
+
 import { Outlet} from "react-router-dom"
 import styles from "./AppLayout.module.scss"
 import Testing from "assets/svg/testing.svg"
 import Logger from "assets/svg/logger.svg"
+import { Navbar } from "components/Navbar/Navbar"
 
 export const AppLayout = () => {
 
     return (
         <div className={styles.appLayout}>
-            <Sidebar items={[
+            <Navbar items={[
                 {
                     path: "/",
                     icon: Testing
