--- ethernet-view/src/styles/App.css
@@ -1,4 +1,0 @@
-.App {
-  width: 100%;
-  height: 100%;
-}
