+++ ethernet-view/src/styles/styles.scss
@@ -5,7 +5,7 @@ $background-color: #dce3eb;
 $title-color: #45677d;
 $normal-text-color: #373c43;
 $alternate-text-color: #505868;
-$orange: hsl(29, 88%, 57%);
+$orange: rgb(242, 142, 49);
 $blue: #317ae7;
 $base-color: hsl(212, 27%, 55%);
